#618 Moved Helper.java to core, and updated data-sly-use in page.html. Refactored Helper.java to allow siteCSS to exist under per:Site/jcr:content
