+++ pagerenderer/server/core/pom.xml
@@ -151,6 +151,13 @@
             <groupId>javax.inject</groupId>
             <artifactId>javax.inject</artifactId>
         </dependency>
+
+        <dependency>
+            <groupId>org.apache.sling</groupId>
+            <artifactId>org.apache.sling.scripting.sightly.compiler.java</artifactId>
+            <version>1.0.18</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 
     <profiles>
